requirements-build.txt
# Build-time only: package app into executables
pyinstaller>=6.0.0
pillow>=10.0.0
